Fix deannotate
