+++ src/ansiblelint/constants.py
@@ -122,7 +122,7 @@ NESTED_TASK_KEYS = [
 SKIPPED_RULES_KEY = "__skipped_rules__"
 LINE_NUMBER_KEY = "__line__"
 FILENAME_KEY = "__file__"
-
+ANNOTATION_KEYS = [FILENAME_KEY, LINE_NUMBER_KEY, SKIPPED_RULES_KEY]
 
 INCLUSION_ACTION_NAMES = {
     "include",
