+++ src/ansiblelint/yaml_utils.py
@@ -41,6 +41,7 @@ from ansiblelint.constants import (
 from ansiblelint.errors import MatchError
 from ansiblelint.file_utils import Lintable
 from ansiblelint.utils import get_action_tasks, normalize_task
+from ansiblelint.constants import ANNOTATION_KEYS
 
 if TYPE_CHECKING:
     # noinspection PyProtectedMember
@@ -67,6 +68,21 @@ rules:
     min-spaces-inside: 0  # yamllint defaults to 0
     max-spaces-inside: 1  # yamllint defaults to 0
 """
+
+
+def deannotate(data: Any) -> Any:
+    """Remove our annotations like __file__ and __line__ and return a JSON serializable object."""
+    if isinstance(data, dict):
+        result = data.copy()
+        for key, value in data.items():
+            if key in ANNOTATION_KEYS:
+                del result[key]
+            else:
+                result[key] = deannotate(value)
+        return result
+    if isinstance(data, list):
+        return [deannotate(item) for item in data if item not in ANNOTATION_KEYS]
+    return data
 
 
 @functools.lru_cache(maxsize=1)
